Make requested PR changes
