+++ web/client/src/components/EventsMap/index.js
@@ -1,14 +1,5 @@
 import React, { Component } from 'react'
 import mapboxgl from 'mapbox-gl'
-import { Dropdown } from 'carbon-components-react'
-import EarthquakeList from '../EarthquakeList'
-
-const items = [
-  { id: 'timeframe-past-7', text: 'Past 7 Days' },
-  { id: 'timeframe-past-30', text: 'Past 30 Days' },
-  { id: 'timeframe-past-90', text: 'Past 90 Days' },
-  { id: 'timeframe-past-year', text: 'Past Year' },
-]
 
 console.log(process.env.REACT_APP_MAPBOX_ACCESS_TOKEN)
 
@@ -62,27 +53,6 @@ export default class EventsMap extends Component {
     return (
       <>
         <div ref={(el) => (this.mapWrapper = el)} className="map-wrapper" />
-        <div className="events-panel">
-          <Dropdown
-            id="timeperiod-dropdown"
-            items={items}
-            itemToString={(item) => (item ? item.text : '')}
-            label={items[0].text}
-          />
-          <div className="events-panel__information">
-            <div>
-              <h4>8</h4>
-              <span>Earthquakes detected</span>
-            </div>
-            <div>
-              <h4>2,400</h4>
-              <span>Subscribers alerted</span>
-            </div>
-          </div>
-          <div className="events-page-earthquake-list__container">
-            <EarthquakeList />
-          </div>
-        </div>
       </>
     )
   }
